backend/requirements.txt
Flask==2.3.2
Flask-Cors==3.0.10
Flask-SQLAlchemy==3.0.3
mysql-connector-python==8.1.0
python-dotenv==1.0.0
Flask
Flask-CORS
Flask-SQLAlchemy
mysql-connector-python
Werkzeug
